requirements.txt
# Core dependencies
albumentations>=1.3.0
numpy>=1.20.0
pandas>=1.3.0
pydicom>=2.2.0
tqdm>=4.60.0
nibabel>=3.2.1
scikit-image>=0.18.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
torch>=1.10.0
torchvision>=0.11.0

# For data validation and schemas
jsonschema>=4.0.0
pyyaml>=6.0.0

# For API and server
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6

# For optional functionality
polars>=0.13.0  # Optional, for faster data processing
